Add TPROXY support for OpenBSD
This adds support for the OpenBSD 'divert' target in PF which presents
the client and remote IPs directly to Squid in accept() parameters
the way Linux TPROXY target does.

It also adds support for the SO_BINDANY option on outgoing traffic for
client IP address spoofing which completes teh TPROXY behaviour.

To enable these features Squid is configured the same as Linux TPROXY:
  http_port 1234 tproxy