Performance: pass size details to HttpHeader parse logic
The Http1::RequestParser class is already aware of the location and size
of the mime-header block so there is no need to scan for it again.

This removes the need for httpMsgIsolateHeaders() outside of HttpMsg and
the 1-2 data scans it performs over the mime-header block per HTTP
request.

Also, supporting changes required in HTCP to use the new
HttpHeader::parseHeaders API removes 10x strlen() calls over the HTCP
mime-headers payload and similar large memory blocks per HTCP packet.