SMP shared cache with timeouts for squid
This patch investigates the Ipc::MemMap class which is a shared cache with 
timeouts for use with squid SMP.

TODO: Ipc::MemMap class has similar interfaces and functionality with the 
Ipc::StoreMap class. These two classes should implemented as kid classes of
an Ipc::SharedCache class which encompass features from Ipc::MemMap and
Ipc::StoreMap classes.

This is a Measurement Factory project