Avoid comm.cc:170: "Comm::IsConnOpen(conn)" assertions by being more careful with the client data connection state.
Do not forward a 125/150 response to the client until the client data
connection is ready. Do not forward it at all if the data connection was
already closed by the time we got a server 125/150 response. If we have to
wait, we do not really forward the original 125/150 response but generate our
own 150. That may change.